# lothrop-tech-task-main/backend/api/types.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for CompanyJurisdiction.
const (
	CompanyJurisdictionCaymens   CompanyJurisdiction = "Caymens"
	CompanyJurisdictionSingapore CompanyJurisdiction = "Singapore"
	CompanyJurisdictionUK        CompanyJurisdiction = "UK"
)

// Defines values for CreateCompanyRequestJurisdiction.
const (
	CreateCompanyRequestJurisdictionCaymens   CreateCompanyRequestJurisdiction = "Caymens"
	CreateCompanyRequestJurisdictionSingapore CreateCompanyRequestJurisdiction = "Singapore"
	CreateCompanyRequestJurisdictionUK        CreateCompanyRequestJurisdiction = "UK"
)

// Defines values for GetCompaniesParamsJurisdiction.
const (
	Caymens   GetCompaniesParamsJurisdiction = "Caymens"
	Singapore GetCompaniesParamsJurisdiction = "Singapore"
	UK        GetCompaniesParamsJurisdiction = "UK"
)

// ApiResponse defines model for ApiResponse.
type ApiResponse struct {
	Error bool   `json:"error"`
	Msg   string `json:"msg"`
}

// CompaniesResponse defines model for CompaniesResponse.
type CompaniesResponse struct {
	Companies []Company `json:"companies"`
	Limit     int       `json:"limit"`
	Offset    int       `json:"offset"`
	Total     int       `json:"total"`
}

// Company defines model for Company.
type Company struct {
	CompanyAddress       string              `json:"company_address"`
	CompanyName          string              `json:"company_name"`
	DateCreated          time.Time           `json:"date_created"`
	DateUpdated          time.Time           `json:"date_updated"`
	Id                   openapi_types.UUID  `json:"id"`
	Jurisdiction         CompanyJurisdiction `json:"jurisdiction"`
	NatureOfBusiness     *string             `json:"nature_of_business"`
	NumberOfDirectors    *int                `json:"number_of_directors"`
	NumberOfShareholders *int                `json:"number_of_shareholders"`
	SecCode              *string             `json:"sec_code"`
}

// CompanyJurisdiction defines model for Company.Jurisdiction.
type CompanyJurisdiction string

// CreateCompanyRequest defines model for CreateCompanyRequest.
type CreateCompanyRequest struct {
	CompanyAddress       string                           `json:"company_address"`
	CompanyName          string                           `json:"company_name"`
	Jurisdiction         CreateCompanyRequestJurisdiction `json:"jurisdiction"`
	NatureOfBusiness     *string                          `json:"nature_of_business"`
	NumberOfDirectors    *int                             `json:"number_of_directors"`
	NumberOfShareholders *int                             `json:"number_of_shareholders"`
	SecCode              *string                          `json:"sec_code"`
}

// CreateCompanyRequestJurisdiction defines model for CreateCompanyRequest.Jurisdiction.
type CreateCompanyRequestJurisdiction string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error bool   `json:"error"`
	Msg   string `json:"msg"`
}

// GetCompaniesParams defines parameters for GetCompanies.
type GetCompaniesParams struct {
	// Limit Maximum number of companies to return
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`

	// Offset Number of companies to skip for pagination
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`

	// Jurisdiction Filter companies by jurisdiction
	Jurisdiction *GetCompaniesParamsJurisdiction `form:"jurisdiction,omitempty" json:"jurisdiction,omitempty"`
}

// GetCompaniesParamsJurisdiction defines parameters for GetCompanies.
type GetCompaniesParamsJurisdiction string

// CreateCompanyJSONRequestBody defines body for CreateCompany for application/json ContentType.
type CreateCompanyJSONRequestBody = CreateCompanyRequest